Rename template package to component
