+++ ui/component/doc.go
@@ -1,4 +1,4 @@
-// Package template features a framework, inspired by React and Redux,
+// Package component features a framework, inspired by React and Redux,
 // that allows you to extend the basic ui package with custom Components
 // on top of plain ui Elements.
-package template
+package component
