+++ ui/component/state.go
@@ -1,4 +1,4 @@
-package template
+package component
 
 import (
 	"fmt"
